requirements.txt
gremlinpython==3.5.0
pytest
python-igraph
cairocffi # for visualization of igraph